+++ js/util.js
@@ -1,0 +1,35 @@
+// Задание считается выполненным, если в проекте описаны следующие функции:
+// Функция, возвращающая случайное число с плавающей точкой из переданного диапазона включительно.
+const getRandomPositiveFloat = (a, b, digits = 1) => {
+  if (a < 0 || b < 0 || digits < 0) {
+    return NaN;
+  }
+  // Чтобы не заставлять пользователя нашей функции помнить порядок аргументов,
+  // реализуем поддержку передачи минимального и максимального значения в любом порядке,
+  // а какое из них большее и меньшее вычислим с помощью Math.min и Math.max
+  const lower = Math.min(a, b);
+  const upper = Math.max(a, b);
+  // Дальше используем Math.random() для получения случайного дробного числа в диапазоне [0, 1),
+  // которое домножаем на разницу между переданными числами - это будет наша случайная дельта.
+  // После нужно сложить дельту с минимальным значением, чтобы получить итоговое случайное число.
+  const result = Math.random() * (upper - lower) + lower;
+  // И в конце с помощью метода toFixed любого числа в JavaScript
+  // указать требуемое количество знаков после точки.
+  // Метод возвращает строку, поэтому с помощью унарного плюса превращаем её в число
+  return +result.toFixed(digits);
+};
+
+// Функция, возвращающая случайное целое число из переданного диапазона включительно.
+const getRandomPositiveInteger = (a, b) => {
+  if (a < 0 || b < 0 || a === b) {
+    return NaN;
+  }
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+export {getRandomPositiveFloat, getRandomPositiveInteger, getRandomArrayElement};
